unity-cli@v1.2.2 (#24)
- added additional version checks
- add Rosetta 2 exec on mac for x86_64 editors
- add support for floating license config as path AND base64 string
- fix occasional race condition when starting editor process exits faster than logs can be seen

+++ src/license-client.ts
@@ -354,8 +354,19 @@ export class LicensingClient {
                         throw new Error(`Unsupported platform: ${process.platform}`);
                 }
 
+                // Ensure the services directory exists
+                if (!fs.existsSync(servicesPath)) {
+                    await fs.promises.mkdir(servicesPath, { recursive: true });
+                }
+
                 const servicesConfigPath = path.join(servicesPath, 'services-config.json');
-                await fs.promises.writeFile(servicesConfigPath, Buffer.from(options.servicesConfig, 'base64'));
+
+                if (fs.existsSync(options.servicesConfig)) {
+                    await fs.promises.copyFile(options.servicesConfig, servicesConfigPath);
+                }
+                else {
+                    await fs.promises.writeFile(servicesConfigPath, Buffer.from(options.servicesConfig, 'base64'));
+                }
                 break;
             }
             default: { // personal and professional license activation
